+++ lib/controllers/application.rb
@@ -1,7 +1,7 @@
 require_relative '../../lib/views/observable'
-require_relative '../../lib/views/windows/main_window'
+require_relative '../../lib/views/windows/app_window'
 require_relative '../../lib/views/windows/main_menu_window'
-require_relative '../../lib/views/windows/game_window_widget'
+require_relative '../../lib/views/windows/game_window'
 
 module C4
   class Application < Gtk::Application
@@ -13,8 +13,8 @@ module C4
       super 'com.rubynetix.connect4', Gio::ApplicationFlags::FLAGS_NONE
 
       signal_connect :activate do |application|
-        windows = [MainMenuWindow.new, GameWindowWidget.new]
-        @ui = C4::MainWindow.new(application, windows[0].id, windows)
+        windows = [MainMenuWindow.new, GameWindow.new]
+        @ui = C4::AppWindow.new(application, windows[0].id, windows)
 
         # Listen for events in spawned windows and bubble them up the observable
         # chain to the wrapping GtkUI instance
