+++ lib/resources/ui/app_window.ui
@@ -2,7 +2,7 @@
 <!-- Generated with glade 3.22.1 -->
 <interface>
   <requires lib="gtk+" version="3.20"/>
-  <template class="C4MainWindow" parent="GtkApplicationWindow">
+  <template class="C4AppWindow" parent="GtkApplicationWindow">
     <property name="can_focus">False</property>
     <child>
       <placeholder/>
