+++ lib/resources/ui/main_menu_window.ui
